python-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
httpx==0.27.2
pillow==10.2.0
opencv-python==4.9.0.80
pytesseract==0.3.10
google-cloud-vision==3.5.0
numpy==1.26.3
python-dotenv==1.0.0
aiofiles==23.2.1
